Changes views, controllers, tests to respect new routing
