+++ app/components/candidate_interface/degree_empty_component.html.erb
@@ -4,4 +4,4 @@
 
 <p class="govuk-body">You can also add any other degrees that you have, or you’re working towards.</p>
 
-<%= govuk_button_link_to degrees.empty? ? t('application_form.degree.add.button') : t('application_form.degree.another.button'), candidate_interface_new_degree_country_path, primary: true %>
+<%= govuk_button_link_to degrees.empty? ? t('application_form.degree.add.button') : t('application_form.degree.another.button'), candidate_interface_degree_country_path, primary: true %>
