+++ app/components/candidate_interface/degree_review_component.html.erb
@@ -3,7 +3,7 @@
     <%= render(SummaryCardHeaderComponent.new(title: title(degree), heading_level: @heading_level)) do %>
       <% if @editable %>
         <div class="app-summary-card__actions">
-          <%= govuk_link_to candidate_interface_new_confirm_degree_destroy_path(degree.id) do %>
+          <%= govuk_link_to candidate_interface_confirm_degree_destroy_path(degree.id) do %>
             <%= t('application_form.degree.delete') %><span class="govuk-visually-hidden"><%= generate_action(degree: degree) %></span>
           <% end %>
         </div>
@@ -13,5 +13,5 @@
 <% end %>
 
 <% if show_missing_banner? %>
-  <%= render(CandidateInterface::IncompleteSectionComponent.new(section: :degrees, section_path: candidate_interface_new_degree_review_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::IncompleteSectionComponent.new(section: :degrees, section_path: candidate_interface_degree_review_path, error: @missing_error)) %>
 <% end %>
