+++ app/views/candidate_interface/degrees/degree/new_award_year.erb
@@ -2,7 +2,7 @@
   <div class="govuk-grid-column-two-thirds">
   <% content_for :title, title_with_error_prefix(t("application_form.degree.award_year.new_label", did_or_will: "#{@wizard.completed? ? 'did' : 'will' }"), @wizard.errors.any?) %>
   <% content_for :before_content, govuk_back_link_to(@wizard.award_year_back_link) %>
-  <%= form_with model: @wizard, url: candidate_interface_new_degree_award_year_path do |f| %>
+  <%= form_with model: @wizard, url: candidate_interface_degree_award_year_path do |f| %>
     <%= f.govuk_error_summary %>
     <%= f.govuk_text_field :award_year, label: { text: t("application_form.degree.award_year.new_label", did_or_will: "#{@wizard.completed? ? 'did' : 'will' }"), size: 'l'}, hint: { text: t("application_form.degree.award_year.new_hint") }, width: 4  %>
     <%= f.govuk_submit t('save_and_continue') %>
