+++ app/views/candidate_interface/degrees/degree/new_completed.html.erb
@@ -1,8 +1,8 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
   <% content_for :title, title_with_error_prefix(t('page_titles.degree_completion_status'), @wizard.errors.any?) %>
-  <% content_for :before_content, govuk_back_link_to(@wizard.reviewing_and_unchanged_country? ? @wizard.back_to_review : candidate_interface_new_degree_university_path) %>
-  <%= form_with model: @wizard, url: candidate_interface_new_degree_completed_path do |f| %>
+  <% content_for :before_content, govuk_back_link_to(@wizard.reviewing_and_unchanged_country? ? @wizard.back_to_review : candidate_interface_degree_university_path) %>
+  <%= form_with model: @wizard, url: candidate_interface_degree_completed_path do |f| %>
     <%= f.govuk_error_summary %>
     <%= f.govuk_radio_buttons_fieldset :completed, legend: { text: t('page_titles.degree_completion_status'), size: 'l' } do %>
       <%= f.hidden_field(:completed) %>
