+++ app/views/candidate_interface/degrees/degree/new_enic.html.erb
@@ -1,6 +1,6 @@
 <% content_for :title, title_with_error_prefix(t('page_titles.degree_enic'), @wizard.errors.any?) %>
 <%= content_for :before_content, govuk_back_link_to(@wizard.enic_back_link) %>
-<%= form_with model: @wizard, url: candidate_interface_new_degree_enic_path do |f| %>
+<%= form_with model: @wizard, url: candidate_interface_degree_enic_path do |f| %>
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
     <%= f.govuk_error_summary %>
