+++ app/views/candidate_interface/degrees/degree/new_grade.html.erb
@@ -1,6 +1,6 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= content_for :before_content, govuk_back_link_to(@wizard.reviewing_and_unchanged_country? ? @wizard.back_to_review : candidate_interface_new_degree_completed_path) %>
+    <%= content_for :before_content, govuk_back_link_to(@wizard.reviewing_and_unchanged_country? ? @wizard.back_to_review : candidate_interface_degree_completed_path) %>
     <%= render CandidateInterface::DegreeGradeComponent.new(model: @wizard) %>
   </div>
 </div>
