+++ app/views/candidate_interface/degrees/degree/new_start_year.html.erb
@@ -1,8 +1,8 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
   <% content_for :title, title_with_error_prefix(t('page_titles.what_year_did_you_start_your_degree'), @wizard.errors.any?) %>
-  <% content_for :before_content, govuk_back_link_to(@wizard.reviewing_and_unchanged_country? ? @wizard.back_to_review : candidate_interface_new_degree_grade_path) %>
-  <%= form_with model: @wizard, url: candidate_interface_new_degree_start_year_path do |f| %>
+  <% content_for :before_content, govuk_back_link_to(@wizard.reviewing_and_unchanged_country? ? @wizard.back_to_review : candidate_interface_degree_grade_path) %>
+  <%= form_with model: @wizard, url: candidate_interface_degree_start_year_path do |f| %>
     <%= f.govuk_error_summary %>
     <%= f.govuk_text_field :start_year, label: { text: t('page_titles.what_year_did_you_start_your_degree'), size: 'l' }, hint: { text: 'For example, 1999' }, width: 4 %>
     <%= f.govuk_submit t('save_and_continue') %>
