+++ app/views/candidate_interface/degrees/degree/new_subject.html.erb
@@ -3,7 +3,7 @@
   <% content_for :title, title_with_error_prefix(t('page_titles.degree_subject'), @wizard.errors.any?) %>
   <% content_for :before_content, govuk_back_link_to(@wizard.subject_back_link) %>
 
-  <%= form_with model: @wizard, url: candidate_interface_new_degree_subject_path do |f| %>
+  <%= form_with model: @wizard, url: candidate_interface_degree_subject_path do |f| %>
     <%= f.govuk_error_summary %>
     <%= render DfE::Autocomplete::View.new(
       f,
