+++ app/views/candidate_interface/degrees/degree/new_university.html.erb
@@ -2,7 +2,7 @@
   <div class="govuk-grid-column-two-thirds">
   <% content_for :title, title_with_error_prefix(t('page_titles.degree_university'), @wizard.errors.any?) %>
   <% content_for :before_content, govuk_back_link_to(@wizard.university_back_link) %>
-  <%= form_with model: @wizard, url: candidate_interface_new_degree_university_path do |f| %>
+  <%= form_with model: @wizard, url: candidate_interface_degree_university_path do |f| %>
     <%= f.govuk_error_summary %>
     <%= f.govuk_fieldset legend: { text: t('page_titles.degree_university'), size: 'l' } do %>
       <% if @wizard.uk? %>
