+++ app/views/candidate_interface/degrees/destroy/confirm_destroy.html.erb
@@ -1,9 +1,9 @@
 <% content_for :title, t('page_titles.destroy_degree') %>
-<% content_for :before_content, govuk_back_link_to(candidate_interface_new_degree_review_path, 'Back') %>
+<% content_for :before_content, govuk_back_link_to(candidate_interface_degree_review_path, 'Back') %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= form_with model: @degree, url: candidate_interface_new_confirm_degree_destroy_path(@degree.id), method: :delete do |f| %>
+    <%= form_with model: @degree, url: candidate_interface_confirm_degree_destroy_path(@degree.id), method: :delete do |f| %>
       <h1 class="govuk-heading-xl">
         <span class="govuk-caption-xl"><%= "#{@degree.qualification_type} #{@degree.subject}" %></span>
         <%= t('page_titles.destroy_degree') %>
@@ -12,7 +12,7 @@
       <%= f.govuk_submit t('application_form.degree.confirm_delete'), warning: true %>
 
       <p class="govuk-body">
-        <%= govuk_link_to 'Cancel', candidate_interface_new_degree_review_path %>
+        <%= govuk_link_to 'Cancel', candidate_interface_degree_review_path %>
       </p>
     <% end %>
   </div>
