+++ app/views/candidate_interface/degrees/review/show.html.erb
@@ -1,7 +1,7 @@
 <% content_for :title, t('page_titles.degree') %>
 <% content_for :before_content, govuk_back_link_to(candidate_interface_application_form_path, 'Back to application') %>
 
-<%= form_with model: @section_complete_form, url: candidate_interface_new_degree_complete_path, method: :patch do |f| %>
+<%= form_with model: @section_complete_form, url: candidate_interface_degree_complete_path, method: :patch do |f| %>
   <%= f.govuk_error_summary %>
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
@@ -11,7 +11,7 @@
 
       <% degree_empty_component = CandidateInterface::DegreeEmptyComponent.new(application_form: @application_form) %>
       <% unless degree_empty_component.render? %>
-        <%= govuk_button_link_to t('application_form.degree.another.button'), candidate_interface_new_degree_country_path(context: :new_degree), secondary: true %>
+        <%= govuk_button_link_to t('application_form.degree.another.button'), candidate_interface_degree_country_path(context: :new_degree), secondary: true %>
       <% end %>
     </div>
   </div>
